refactor(common): unite PendingClass with ApiContractClass
